requirements-gpu.txt
tensorflow-gpu==2.2.0
opencv-python==4.1.1.26
lxml
tqdm

-e .
